backend: CORS, round 3
I misunderstood the preflight CORS check, and what Chrome's devtools were
telling me after flailing about debugging for a minute. This properly adds an
OPTIONS endpoint that can vouch for the cross-domain request from the frontend.

As usual, I totally haven't tested this, but it compiles. So we'll see if round
3 is a KO or not.

Signed-off-by: Austin Seipp <aseipp@pobox.com>