front/back: refactor pubkey.js into data.js
Now, there's another new environment variable, CHARGE_DOMAIN, which is specified
to be whatever domain is hosting this service, with a /charge endpoint. This
means that the donate.js file can be completely ignorant of the submission
domain; only donate.html needs to properly include data.js (before donate.js),
and it will have the stripe key, AND the domain to submit the POST request to.

This makes testing a little easier.

Signed-off-by: Austin Seipp <aseipp@pobox.com>